cambios para obtener porcentajes por id de factura

+++ src/main/java/ar/my/mensualidades/dto/PorcentajeUsuariosFacturasDto.java
@@ -13,5 +13,5 @@ public class PorcentajeUsuariosFacturasDto extends Dto implements Serializable {
     private final Long idFactura;
     private final Long idUsuario;
     private final String usuario;
-    private final Float porcentaje;
+    private final Double porcentaje;
 }
